separated the main file from the map list widget
